requirements.txt
mcp>=1.0.0
openai>=1.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
typing-extensions>=4.0.0
pydantic>=2.0.0
aiohttp>=3.9.0 